# bookshelf-microservice/src/main/resources/API-specification.yaml
openapi: 3.0.3
info:
  title: Bookshelf Microservice (Team 18a)
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/bookshelf_service'
    description: Local server for team 18a
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Create new user
      description: Create a new user, this is done to keep everything in sync with the user microservice
      requestBody:
        content:
          application/json:
            schema:
              description: ID of the user.
              type: string
              format: uuid
              example: "b123e456-7890-1234-5678-9abcdef01234"
      responses:
        '200':
          description: Successful response.
        '400':
          description: Invalid request.
        '500':
          description: Internal server error.

    delete:
      tags:
        - user
      summary: Deletes a user
      description: Deletes a user, all of things tied to the user such as shelves, statistics etc will be deleted.
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of the user that needs to be deleted.
          schema:
            type: string
            format: uuid
            example: "b123e456-7890-1234-5678-9abcdef01234"
      responses:
        '200':
          description: Successful response.
        '400':
          description: Invalid request.
        '404':
          description: User not found.
        '500':
          description: Internal server error.
...